Update discord.js and npm-check-updates libraries.
Some sort of Discord API change has resulted in the following errors at
the end of the `/dota` timer:
Error: Expected token to be set for this request, but none was present

The fix is to update the `discord.js` library to the latest version. The
updated `discord.js` library issues warnings about newly deprecated
interaction response options: `fetchReply` and `ephemeral`. Those have
been replaced with `withResponse` and `flags`, respectively.

This also updates the `npm-check-updates` library.

The `openai` library is left as-is since upgrading from 3.2.1 to 5.9.0
would be a breaking change.